requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
sqlalchemy>=2.0
pydantic>=2.6
requests>=2.31
sentence-transformers>=2.6
google-genai
python-dotenv